# openapi.yaml
openapi: 3.0.3
info:
  title: API Tablero Cooptech
  version: 1.0.0
  description: API REST del Tablero de Mando Cooptech. Single-tenant, uso interno.
    Autenticación delegada al servicio de identidad central (Bearer JWT).
servers:
- url: /api/v1
security:
- bearerAuth: []
...
paths:
  /colaboradores:
    get:
      tags:
      - colaboradores
      summary: Listar colaboradores
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Colaborador'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - colaboradores
      summary: Crear colaboradore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ColaboradorInput'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Colaborador'
        '422':
          $ref: '#/components/responses/Validation'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /colaboradores/{id}:
    parameters:
    - $ref: '#/components/parameters/Id'
    get:
      tags:
      - colaboradores
      summary: Obtener por id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Colaborador'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
      - colaboradores
      summary: Actualizar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ColaboradorInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Colaborador'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Validation'
    delete:
      tags:
      - colaboradores
      summary: Eliminar
      responses:
        '204':
          description: Eliminado
        '404':
          $ref: '#/components/responses/NotFound'
  /proyectos:
    get:
      tags:
      - proyectos
      summary: Listar proyectos
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Proyecto'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - proyectos
      summary: Crear proyecto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProyectoInput'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proyecto'
        '422':
          $ref: '#/components/responses/Validation'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /tareas:
    get:
      tags:
      - tareas
      summary: Listar tareas
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tarea'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - tareas
      summary: Crear tarea
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TareaInput'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tarea'
        '422':
          $ref: '#/components/responses/Validation'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /objetivos:
    get:
      tags:
      - objetivos
      summary: Listar objetivos
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Objetivo'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - objetivos
      summary: Crear objetivo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObjetivoInput'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Objetivo'
        '422':
          $ref: '#/components/responses/Validation'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /tags:
    get:
      tags:
      - tags
      summary: Listar tags
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tag'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - tags
      summary: Crear tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagInput'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '422':
          $ref: '#/components/responses/Validation'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /leads:
    get:
      tags:
      - leads
      summary: Listar leads
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Lead'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - leads
      summary: Crear lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadInput'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
        '422':
          $ref: '#/components/responses/Validation'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /clientes:
    get:
      tags:
      - clientes
      summary: Listar clientes
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Cliente'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - clientes
      summary: Crear cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClienteInput'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
        '422':
          $ref: '#/components/responses/Validation'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /guardias:
    get:
      tags:
      - guardias
      summary: Listar guardias
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Guardia'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - guardias
      summary: Crear guardia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuardiaInput'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guardia'
        '422':
          $ref: '#/components/responses/Validation'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /francos:
    get:
      tags:
      - francos
      summary: Listar francos
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FrancoEspecial'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - francos
      summary: Crear franco
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FrancoEspecialInput'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrancoEspecial'
        '422':
          $ref: '#/components/responses/Validation'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /feriados:
    get:
      tags:
      - feriados
      summary: Listar feriados
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Feriado'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - feriados
      summary: Crear feriado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeriadoInput'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feriado'
        '422':
          $ref: '#/components/responses/Validation'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /colaboradores/{id}/periodos:
    parameters:
    - $ref: '#/components/parameters/Id'
    get:
      tags:
      - colaboradores
      summary: Períodos de actividad del colaborador
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Periodo'
    post:
      tags:
      - colaboradores
      summary: Agregar período
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PeriodoInput'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Periodo'
...
  /archivos:
    post:
      tags:
      - archivos
      summary: Subir archivo a MinIO
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                leadId: &id002
                  type: integer
                contexto: &id004
                  type: string
                gpsLat: &id001
                  type: number
                gpsLng: *id001
                esBoceto: &id005
                  type: boolean
      responses:
        '201':
          description: Subido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Archivo'
...
  /grilla:
    get:
      tags:
      - grilla
      summary: Entradas de grilla por rango/colaborador
      parameters:
      - name: colaboradorId
        in: query
        schema: *id002
      - name: desde
        in: query
        schema: &id003
          type: string
          format: date
      - name: hasta
        in: query
        schema: *id003
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GrillaEntrada'
    put:
      tags:
      - grilla
      summary: Crear o actualizar la entrada de un día (upsert por colaborador+fecha)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrillaEntradaInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrillaEntrada'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: *id002
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 50
  responses:
    Unauthorized:
      description: Token ausente o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: No encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Validation:
      description: Datos inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: *id004
            message: *id004
    Pagination:
      type: object
      properties:
        page: *id002
        pageSize: *id002
        total: *id002
    Colaborador:
      type: object
      properties:
        id: *id002
        identitySub: &id006
          type: string
          nullable: true
        tipo: &id007
          type: string
          enum:
          - manager
          - gerencial
          - colaborador
          - otras_areas
          - tercerizado
        nombre: *id004
        email: &id008
          type: string
          nullable: true
        sector: &id009
          type: string
          nullable: true
        iniciales: &id010
          type: string
          nullable: true
        haceGuardia: *id005
        fechaIngreso: &id011
          type: string
          format: date
          nullable: true
        cumpleDia: &id012
          type: integer
          nullable: true
        cumpleMes: &id013
          type: integer
          nullable: true
        activo: *id005
    ColaboradorInput:
      type: object
      properties:
        identitySub: *id006
        tipo: *id007
        nombre: *id004
        email: *id008
        sector: *id009
        iniciales: *id010
        haceGuardia: *id005
        fechaIngreso: *id011
        cumpleDia: *id012
        cumpleMes: *id013
        activo: *id005
    Proyecto:
      type: object
      properties:
        id: *id002
        nombre: *id004
        enfoque: &id014
          type: string
          enum: &id023
          - Organización
          - Electrónica
          - Desarrollo web
          - Comercialización
          - Operación
          nullable: true
        estado: &id015
          type: string
          enum:
          - activo
          - pausado
          - cerrado
        leadId: &id016
          type: integer
          nullable: true
    ProyectoInput:
      type: object
      properties:
        nombre: *id004
        enfoque: *id014
        estado: *id015
        leadId: *id016
    Tarea:
      type: object
      properties:
        id: *id002
        proyectoId: &id017
          type: integer
          nullable: true
        titulo: *id004
        descripcion: &id018
          type: string
          nullable: true
        kanbanCol: &id019
          type: string
          enum:
          - backlog
          - todo
          - doing
          - done
        prioridad: &id020
          type: string
          enum:
          - baja
          - media
          - alta
          - urgente
        orden: *id002
        responsablesIds: &id021
          type: array
          items: *id002
        tagsIds: &id022
          type: array
          items: *id002
    TareaInput:
      type: object
      properties:
        proyectoId: *id017
        titulo: *id004
        descripcion: *id018
        kanbanCol: *id019
        prioridad: *id020
        orden: *id002
        responsablesIds: *id021
        tagsIds: *id022
    Objetivo:
      type: object
      properties:
        id: *id002
        codigo: *id004
        titulo: *id004
        descripcion: &id024
          type: string
          nullable: true
        indicador: &id025
          type: string
          nullable: true
        meta: &id026
          type: string
          nullable: true
        peso: &id027
          type: number
          nullable: true
        fechaEsperada: &id028
          type: string
          format: date
          nullable: true
        enfoque: &id029
          type: string
          enum: *id023
          nullable: true
        calculo: &id030
          type: string
          enum:
          - manual
          - por_tags
          - por_actividad
        tagsIds: &id031
          type: array
          items: *id002
    ObjetivoInput:
      type: object
      properties:
        codigo: *id004
        titulo: *id004
        descripcion: *id024
        indicador: *id025
        meta: *id026
        peso: *id027
        fechaEsperada: *id028
        enfoque: *id029
        calculo: *id030
        tagsIds: *id031
    Tag:
      type: object
      properties:
        id: *id002
        nombre: *id004
        categoria: &id032
          type: string
          nullable: true
        color: &id033
          type: string
          nullable: true
    TagInput:
      type: object
      properties:
        nombre: *id004
        categoria: *id032
        color: *id033
    Lead:
      type: object
      properties:
        id: *id002
        organizacion: *id004
        contactoNombre: &id035
          type: string
          nullable: true
        telefono: &id036
          type: string
          nullable: true
        email: &id037
          type: string
          nullable: true
        ciudad: &id038
          type: string
          nullable: true
        fechaPrimerContacto: &id039
          type: string
          format: date
          nullable: true
        ownerId: &id040
          type: integer
          nullable: true
        etapa: &id041
          type: string
          enum:
          - contacto
          - visita_agendada
          - visita_realizada
          - propuesta
          - negociacion
          - trial
          - ganado
          - perdido
        valorEstimadoUsd: &id042
          type: number
          nullable: true
        montoFacturadoUsd: &id043
          type: number
          nullable: true
        productos: &id044
          type: array
          items: *id004
        presupuestoEstado: &id034
          type: object
          nullable: true
        presupuestoAguaEstado: *id034
        coopcloudEstado: *id034
        coopcloudCostoMensual: &id045
          type: number
          nullable: true
        notas: &id046
          type: string
          nullable: true
        fuente: &id047
          type: string
          nullable: true
    LeadInput:
      type: object
      properties:
        organizacion: *id004
        contactoNombre: *id035
        telefono: *id036
        email: *id037
        ciudad: *id038
        fechaPrimerContacto: *id039
        ownerId: *id040
        etapa: *id041
        valorEstimadoUsd: *id042
        montoFacturadoUsd: *id043
        productos: *id044
        presupuestoEstado: *id034
        presupuestoAguaEstado: *id034
        coopcloudEstado: *id034
        coopcloudCostoMensual: *id045
        notas: *id046
        fuente: *id047
    Cliente:
      type: object
      properties:
        id: *id002
        nombre: *id004
        tipoCliente: &id048
          type: string
          nullable: true
    ClienteInput:
      type: object
      properties:
        nombre: *id004
        tipoCliente: *id048
    Guardia:
      type: object
      properties:
        id: *id002
        colaboradorId: *id002
        fechaInicio: *id003
        fechaFin: &id049
          type: string
          format: date
          nullable: true
        notas: &id050
          type: string
          nullable: true
    GuardiaInput:
      type: object
      properties:
        colaboradorId: *id002
        fechaInicio: *id003
        fechaFin: *id049
        notas: *id050
    FrancoEspecial:
      type: object
      properties:
        id: *id002
        colaboradorId: *id002
        fecha: *id003
        tipo: &id051
          type: string
          nullable: true
        motivo: &id052
          type: string
          nullable: true
    FrancoEspecialInput:
      type: object
      properties:
        colaboradorId: *id002
        fecha: *id003
        tipo: *id051
        motivo: *id052
    Feriado:
      type: object
      properties:
        id: *id002
        fecha: *id003
        nombre: *id004
    FeriadoInput:
      type: object
      properties:
        fecha: *id003
        nombre: *id004
...
    Periodo:
      type: object
      properties:
        id: *id002
        colaboradorId: *id002
        desde: *id003
        hasta:
          type: string
          format: date
          nullable: true
    PeriodoInput:
      type: object
      properties:
        desde: *id003
        hasta:
          type: string
          format: date
          nullable: true
      required:
      - desde
...
    Archivo:
      type: object
      properties:
        id: *id002
        key: *id004
        url:
          type: string
          nullable: true
        nombre: *id004
        mime:
          type: string
          nullable: true
        tamano:
          type: integer
          nullable: true
        gpsLat:
          type: number
          nullable: true
        gpsLng:
          type: number
          nullable: true
        esBoceto: *id005
        leadId:
          type: integer
          nullable: true
        contexto:
          type: string
          nullable: true
    GrillaEntrada:
      type: object
      properties:
        id: *id002
        colaboradorId: *id002
        fecha: *id003
        estado:
          type: string
          enum:
          - present
          - home_office
          - vacaciones
          - franco
          - franco_cumple
          - feriado
          - licencia
          - viaje
        items:
          type: array
          items:
            type: object
          nullable: true
    GrillaEntradaInput:
      type: object
      properties:
        colaboradorId: *id002
        fecha: *id003
        estado:
          type: string
          enum:
          - present
          - home_office
          - vacaciones
          - franco
          - franco_cumple
          - feriado
          - licencia
          - viaje
        items:
          type: array
          items:
            type: object
          nullable: true
      required:
      - colaboradorId
      - fecha
...